Moves Identity schema creation to packages
